feat: 쿼리 파일 구조 변경 및 CurrentLevelSection 수정
- point.ts 파일을 grade.ts로 변경하여 사용자 등급 관련 쿼리로 통합하였습니다.
- CurrentLevelSection에서 pointQueryOptions를 gradeQueryOptions로 수정하였습니다.
